+++ packages/stream_chat_flutter/lib/src/system_message.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:stream_chat_flutter/src/extension.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
 /// {@macro system_message}
@@ -21,16 +22,22 @@ class StreamSystemMessage extends StatelessWidget {
 
   /// The function called when tapping on the message
   /// when the message is not failed
+  // TODO: Convert it to a simple OnTap instead.
   final void Function(Message)? onMessageTap;
 
   @override
   Widget build(BuildContext context) {
     final theme = StreamChatTheme.of(context);
+    final message = this.message.replaceMentions();
+
+    final messageText = message.text;
+    if (messageText == null) return const SizedBox.shrink();
+
     return GestureDetector(
       behavior: HitTestBehavior.opaque,
       onTap: onMessageTap == null ? null : () => onMessageTap!(message),
       child: Text(
-        message.text!,
+        messageText,
         textAlign: TextAlign.center,
         softWrap: true,
         style: theme.textTheme.captionBold.copyWith(