+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -3,7 +3,10 @@
 🐞 Fixed
 
 - [[#1247]](https://github.com/GetStream/stream-chat-flutter/issues/1247) Fix Jiffy initialization.
-- [[#1232]](https://github.com/getstream/stream-chat-flutter/issues/1232) Fix DateDivider not showing up in the chat.
+- [[#1232]](https://github.com/getstream/stream-chat-flutter/issues/1232) Fix DateDivider not
+  showing up in the chat.
+- [[#1240]](https://github.com/getstream/stream-chat-flutter/issues/1240) Substitute mentioned user
+  ids with user names in system message.
 
 ## 4.4.0